+++ packages/pluggableWidgets/combobox-web/src/helpers/Database/utils.ts
@@ -28,12 +28,13 @@ type ExtractionReturnValue = {
     lazyLoading: boolean;
     loadingType: LoadingTypeEnum;
     valueSourceAttribute: ListAttributeValue<string | Big> | undefined;
+    datasourceFilterDebounceInterval: number;
 };
 
 export function extractDatabaseProps(props: ComboboxContainerProps): ExtractionReturnValue {
     const targetAttribute = props.databaseAttributeString;
     const filterType = props.filterType;
-
+    const datasourceFilterDebounceInterval = props.datasourceFilterDebounceInterval;
     const ds = props.optionsSourceDatabaseDataSource;
     if (!ds) {
         throw new Error("'optionsSourceType' type is 'database' but 'optionsSourceDatabaseDataSource' is not defined.");
@@ -83,7 +84,8 @@ export function extractDatabaseProps(props: ComboboxContainerProps): ExtractionR
         filterType,
         lazyLoading,
         loadingType,
-        valueSourceAttribute
+        valueSourceAttribute,
+        datasourceFilterDebounceInterval
     };
 }