+++ packages/pluggableWidgets/combobox-web/typings/ComboboxProps.d.ts
@@ -100,6 +100,7 @@ export interface ComboboxContainerProps {
     a11yInstructions?: DynamicValue<string>;
     lazyLoading: boolean;
     loadingType: LoadingTypeEnum;
+    datasourceFilterDebounceInterval: number;
     selectedItemsSorting: SelectedItemsSortingEnum;
     filterType: FilterTypeEnum;
 }
@@ -158,6 +159,7 @@ export interface ComboboxPreviewProps {
     a11yInstructions: string;
     lazyLoading: boolean;
     loadingType: LoadingTypeEnum;
+    datasourceFilterDebounceInterval: number | null;
     selectedItemsSorting: SelectedItemsSortingEnum;
     filterType: FilterTypeEnum;
 }