+++ packages/pluggableWidgets/combobox-web/CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - We fixed an issue where combobox lazy load is not working on initial load.
 
+### Added
+
+- We added a configurable debounce interval for datasource filter operations via a new property in the Advanced tab.
+
 ## [2.5.0] - 2025-08-12
 
 ### Added